requirements.txt
bump2version
flake8 < 5
github3.py
pytest
pytest-benchmark
pytest-cov
pytest-mypy
pytest-flake8
pytest-isort
pytest-xdist
python-telegram-bot>=20.0a4
requests
twine
types-requests